Rejected from Acta Astronautica
Dr. Lee had some good advice:

It is unfortunate that the decision was made by only one review. Not sure the AE handled the reviewer process up to the standard. Also, the reviewer’s comments provide several minor observations without stating the key reasons why he/she is against.

We can certainly ask more clarification, but it is likely that it’s going to be a waste of time.

I suggest we resubmit it to the Journal of the Astronautical Sciences, after handling selected comments from the review.